# apiServer/app/api/openapi.yaml
openapi: 3.0.0
info:
  title: Community Model API
  version: v.1.1
  description: >-
    API for accessing data generated by the Community Model, for updating user
    data used by the Community Model for finding communities and for inserting
    new perspectives.
  contact:
    email: gjimenez@ucm.es
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: ''
servers:
  - url: /v1.1/
security:
  - basicAuth: []
...
paths:
  /logs:
    get:
      tags:
        - '[DEV] Community Model Logs'
      parameters:
        - style: form
          explode: true
          name: nLogs
          description: N latest logs. Use '0' to get all logs
          schema:
            minimum: 0
            type: integer
          in: query
          required: true
        - style: form
          explode: true
          name: logsType
          description: Type of logs
          schema:
            default: ALL
            enum:
              - ALL
              - DEBUG
              - INFO
              - WARNING
              - ERROR
              - CRITICAL
            type: string
          in: query
          required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/log'
          description: successful operation
        '400':
          description: error
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getNLatestLogs
      summary: Get n latest logs
      description: Returns the last n documents
      x-swagger-router-controller: Logs
  /logs/dateRange:
    get:
      tags:
        - '[DEV] Community Model Logs'
      parameters:
        - style: form
          explode: true
          name: startDate
          description: 'Start date. ISO 8601 format: YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS'
          schema:
            format: date
            type: string
            example: '2017-07-21T17:32:28Z'
          in: query
          required: true
        - style: form
          explode: true
          name: endDate
          description: >-
            End date. The end date must be greater than the start date. ISO 8601
            format: YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS
          schema:
            format: date
            type: string
            example: '2016-03-04T03:00:00'
          in: query
          required: true
        - style: form
          explode: true
          name: logsType
          description: Type of logs
          schema:
            default: ALL
            enum:
              - ALL
              - DEBUG
              - INFO
              - WARNING
              - ERROR
              - CRITICAL
            type: string
          in: query
          required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/log'
          description: successful operation
        '400':
          description: error
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getLogsBetweenTwoDates
      summary: Get logs between date range
      description: Returns logs between date range
      x-swagger-router-controller: Logs
  /seed:
    get:
      tags:
        - '[DEV] VISIR'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/seedFile'
          description: successful operation
        '400':
          description: invalid seed id
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getSeed
      summary: Get seed file
      description: Returns seed configuration file used by VISIR application.
      x-swagger-router-controller: Visir
  /visir/index:
    get:
      tags:
        - '[DEV] VISIR'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/seedFile'
          description: successful operation
        '400':
          description: invalid seed id
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getIndex
      summary: List of all available visualization files
      description: Returns list of all available visualization files.
      x-swagger-router-controller: Visir
  /perspectives:
    get:
      tags:
        - Perspectives
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/perspective'
                x-content-type: application/json
          description: A list with the perspectives in the model.
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/job'
          description: >-
            Job accepted. Updating CM before returning requested data.
            Redirecting to Jobs Manager
        '400':
          description: Community model does not exist
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getPerspectives
      summary: Get all perspectives in the community model.
      description: >-
        Returns a list of the perspectives in the community model. If the CM
        update is necessary returns a job.
      x-swagger-router-controller: Perspectives
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/perspective'
        required: true
      tags:
        - Perspectives
      responses:
        '204':
          description: Added
        '400':
          description: error
        '401':
          $ref: '#/components/responses/401'
      operationId: postPerspective
      summary: Inserts perspective to the Community Model
      description: ''
      x-swagger-router-controller: Perspectives
...
  /communities/{communityId}:
    get:
      tags:
        - Communities
      parameters:
        - style: simple
          explode: false
          name: communityId
          description: ID of community to return
          schema:
            $ref: '#/components/schemas/id'
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/community'
          description: successful operation
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/job'
          description: >-
            Job accepted. Updating CM before returning requested data.
            Redirecting to Jobs Manager
        '400':
          description: invalid community id
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getCommunityById
      summary: Get community description and explanation
      description: >-
        Returns information about a community. If the CM update is necessary
        returns a job.
      x-swagger-router-controller: Communities
...
  /visir/files/{fileId}:
    get:
      tags:
        - '[DEV] VISIR'
      parameters:
        - style: simple
          explode: false
          name: fileId
          description: Id of the file
          schema:
            $ref: '#/components/schemas/jobId'
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/seedFile'
          description: successful operation
        '400':
          description: invalid seed id
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getFile
      summary: Return the specified visualization file
      description: Returns the requested visualization file.
      x-swagger-router-controller: Visir
    x-last-modified: 1678483487526
  /jobs-manager/jobs/{jobId}:
    get:
      tags:
        - '[DEV] Jobs Manager'
      parameters:
        - style: simple
          explode: false
          name: jobId
          description: Id of the job to monitor
          schema:
            $ref: '#/components/schemas/jobId'
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/jobStarted'
                  - $ref: '#/components/schemas/jobCompleted'
          description: Job status
        '400':
          description: Job does not exist
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getJob
      summary: Job status
      description: >-
        Monitor the status of a job. Shows if the status of the job is finished
        or in progress, and returns data if it is finished. User agent should
        send periodically /Get requests to check the completion status of the
        job.
      x-content-type: application/json
      x-swagger-router-controller: JobsRouter
    x-last-modified: 1678486713587
  /jobs-manager/jobs:
    get:
      tags:
        - '[DEV] Jobs Manager'
      responses:
        '200':
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/jobStarted'
                  - $ref: '#/components/schemas/jobCompleted'
          description: Job status
        '400':
          description: Job does not exist
        '401':
          $ref: '#/components/responses/401'
      security:
        - basicAuth: []
      operationId: getJobs
      summary: Status of all jobs
      description: Returns all current jobs.
      x-content-type: application/json
      x-swagger-router-controller: JobsRouter
components:
  schemas:
    id:
      description: Unique id
      type: string
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
    perspectiveId:
      description: Unique perspective id
      type: string
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
    param:
      description: Param
      type: string
      example: param
    community:
      required:
        - id
        - name
        - perspectiveId
        - users
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        name:
          description: Community name
          type: string
          example: elderly
        perspectiveId:
          description: perspectiveId
          type: string
          example: perspectiveId
        explanations:
          description: >-
            a list of community descriptions (maybe empty). It can be computed
            by the explanation module
          type: array
          items:
            type: string
          example: People whose age is above 65
        users:
          description: A list with the users who belong to the community
          type: array
          items:
            $ref: '#/components/schemas/id'
          example:
            - '23'
            - '24'
      example:
        name: elderly
        id: d290f1ee-6c54-4b01-90e6-d701748f0851
        explanation: People whose age is above 65
        perspectiveId: '101'
        users:
          - '23'
          - '24'
    jobId:
      description: Unique job id
      type: string
      example: 4351
    jobStarted:
      title: jobStarted
      type: object
      properties:
        job:
          title: job
          type: object
          properties:
            path:
              description: path
              type: string
              example: /v1.1/jobs/4351
            jobId:
              $ref: '#/components/schemas/jobId'
            name:
              description: name
              type: string
              example: CM Update
            job-state:
              description: job state
              type: string
              example: STARTED
            job-status:
              description: job status
              type: string
              example: INPROGRESS
            start-time:
              description: job start time
              type: string
              example: INPROGRESS
            time-to-autoremove-job:
              description: time left for autoremove the job
              type: string
              example: INPROGRESS
            data:
              description: The job is not done yet
              type: object
    jobCompleted:
      title: jobCompleted
      type: object
      properties:
        job:
          title: job
          type: object
          properties:
            path:
              description: path
              type: string
              example: /v1.1/jobs/4351
            jobId:
              $ref: '#/components/schemas/jobId'
            name:
              description: name
              type: string
              example: CM Update
            job-state:
              description: job state
              type: string
              example: COMPLETED
            job-status:
              description: job status
              type: string
              example: SUCCESS
            start-time:
              description: job start time
              type: string
              example: INPROGRESS
            time-to-autoremove-job:
              description: time left for autoremove the job
              type: string
              example: INPROGRESS
            data:
              oneOf:
                - $ref: '#/components/schemas/perspective'
                - $ref: '#/components/schemas/community'
                - $ref: '#/components/schemas/listPerspectives'
                - $ref: '#/components/schemas/listCommunities'
                - $ref: '#/components/schemas/perspectiveId'
                - description: perspectiveId perspectiveId
                  type: string
                - description: Visfile
                  type: string
              type: object
    listPerspectives:
      description: List of perspectives
      type: array
      items:
        $ref: '#/components/schemas/perspective'
    listCommunities:
      description: List of communities
      type: array
      items:
        $ref: '#/components/schemas/community'
    job:
      required:
        - path
      type: object
      properties:
        path:
          description: job
          type: string
          example: /v1.1/jobs/{jobId}
    log:
      title: log
      type: object
      properties:
        name:
          type: string
        msg:
          type: string
        args:
          type: array
          items:
            type: string
        levelname:
          type: string
        levelno:
          type: integer
        pathname:
          type: string
        filename:
          type: string
        module:
          type: string
        exc_info:
          nullable: true
          type: string
        exc_text:
          nullable: true
          type: string
        stack_info:
          nullable: true
          type: string
        lineno:
          type: integer
        funcName:
          type: string
        created:
          type: number
        msecs:
          type: integer
        relativeCreated:
          type: number
        thread:
          type: integer
        threadName:
          type: string
        processName:
          type: string
        process:
          type: integer
        username:
          type: string
        time:
          format: date
          type: string
        host:
          type: string
        message:
          type: string
    perspective:
      required:
        - id
        - name
        - algorithm
        - similarity_functions
        - user_attributes
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        name:
          description: Perspective name
          type: string
          example: name
        algorithm:
          title: Algorithm used for this perspective
          type: object
          properties:
            name:
              description: The name of the algorithm
              type: string
              example: optics
            params:
              description: Params
              type: array
              items:
                $ref: '#/components/schemas/param'
              example:
                - paramA
                - paramB
        similarity_functions:
          description: similarity_functions
          type: array
          items:
            $ref: '#/components/schemas/sim_function'
        user_attributes:
          type: array
          items:
            title: user_attributes
            type: object
            properties:
              att_name:
                description: att_name
                type: string
                example: att_name
              att_type:
                description: att_type
                type: string
                example: att_type
    seedFile:
      required:
        - artwork_attributes
        - user_attributes
        - interaction_similarity_functions
      type: object
      properties:
        artwork_attributes:
          type: array
          items:
            type: object
            properties:
              sim_function:
                type: object
                properties:
                  name:
                    type: string
                    example: IconClassSimilarityDAO
                  params:
                    type: array
                    items:
                      type: object
                      properties:
                        att_name:
                          type: string
                          example: Iconclass_subjects_curators
                        att_type:
                          type: string
                          example: String
                  on_attribute:
                    type: object
                    properties:
                      att_name:
                        type: string
                        example: Iconclass_subjects_curators
                      att_type:
                        type: string
                        example: String
        user_attributes:
          type: array
          items:
            type: object
            properties:
              att_name:
                type: string
                example: Age
              att_type:
                type: string
                example: String
        interaction_similarity_functions:
          type: array
          items:
            type: object
            properties:
              sim_function:
                type: object
                properties:
                  name:
                    type: string
                    example: ExtendedPlutchikEmotionSimilarityDAO
                  params:
                    type: array
                    items:
                      type: object
                      properties:
                        att_name:
                          type: string
                          example: Iconclass_subjects_curators
                        att_type:
                          type: string
                          example: String
                  on_attribute:
                    type: object
                    properties:
                      att_name:
                        type: string
                        example: emotions
                      att_type:
                        type: string
                        example: String
                  interaction_object:
                    type: object
                    properties:
                      att_name:
                        type: string
                        example: artworkId
                      att_type:
                        type: string
                        example: String
...
    sim_function:
      title: similarity_function
      type: object
      properties:
        name:
          description: name
          type: string
          example: name
        weight:
          description: weight
          type: number
          example: 2
        params:
          description: Params
          type: array
          items:
            $ref: '#/components/schemas/param'
          example:
            - paramA
            - paramB
        on_attribute:
          title: on_attribute
          type: object
          properties:
            att_name:
              description: att_name
              type: string
              example: att_name
            att_type:
              description: att_type
              type: string
              example: att_type
...
    UnauthorizedError:
      required:
        - message
        - code
      type: object
      properties:
        message:
          type: string
        code:
          maximum: 401
          minimum: 401
          type: integer
      x-last-modified: 1678572397915
  responses:
    '401':
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
      description: Unauthorized access.
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
      description: Use `user` / `pass` as the test credentials
...